Phase 4: Move core server implementation
- Move deployments/llm/llm_server.py → core/server/llm_server.py
- Move deployments/llm/builder_llm_server.py → core/server/builder.py
- Update all imports
