+++ python/ray/llm/_internal/serve/deployments/data_parallel/dp_server.py
@@ -8,10 +8,10 @@ from ray.llm._internal.serve.core.configs.llm_config import LLMConfig
 from ray.llm._internal.serve.deployments.data_parallel.dp_rank_assigner import (
     DPRankAssigner,
 )
-from ray.llm._internal.serve.deployments.llm.builder_llm_server import (
+from ray.llm._internal.serve.core.server.builder import (
     build_llm_deployment,
 )
-from ray.llm._internal.serve.deployments.llm.llm_server import LLMServer
+from ray.llm._internal.serve.core.server.llm_server import LLMServer
 from ray.llm._internal.serve.deployments.routers.router import (
     OpenAiIngress,
     make_fastapi_ingress,
