+++ python/ray/llm/tests/serve/cpu/deployments/llm/multiplex/test_lora_deployment_base_client.py
@@ -7,7 +7,7 @@ from fastapi import HTTPException
 
 from ray import serve
 from ray.llm._internal.serve.core.configs.openai_api_models import ModelCard
-from ray.llm._internal.serve.deployments.llm.llm_server import LLMServer
+from ray.llm._internal.serve.core.server.llm_server import LLMServer
 from ray.llm.tests.serve.mocks.mock_vllm_engine import MockVLLMEngine
 from ray.serve.handle import DeploymentHandle
 from ray.serve.llm import LLMConfig, LoraConfig
