+++ python/ray/llm/tests/serve/cpu/deployments/llm/test_builder_llm_server.py
@@ -6,7 +6,7 @@ from ray import serve
 from ray.llm._internal.serve.core.configs.llm_config import (
     LLMConfig,
 )
-from ray.llm._internal.serve.deployments.llm.builder_llm_server import (
+from ray.llm._internal.serve.core.server.builder import (
     build_llm_deployment,
 )
